requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
distlib==0.3.8
Django==4.2.13
django-allauth==0.62.1
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-crispy-forms==2.1
django-resized==1.0.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
filelock==3.13.3
gunicorn==22.0.0
inflection==0.5.1
packaging==24.0
pillow==10.3.0
platformdirs==4.2.0
psycopg2==2.9.10
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.2.0
virtualenv==20.26.2
whitenoise==4.1.2
